devtools/client/webconsole/new-console-output/test/utils/test_getRepeatId.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for getRepeatId()</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for getRepeatId()</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { getRepeatId } = require("devtools/client/webconsole/new-console-output/utils/messages");

  yield testDuplicateValues();
  yield testDifferentValues();
  yield testDifferentSeverities();
  yield testFalsyValues();

  SimpleTest.finish();

  function testDuplicateValues() {
    const message1 = stubConsoleMessages.get("console.log('foobar', 'test')");
    const message2 = message1.set("repeat", 3);

    is(getRepeatId(message1), getRepeatId(message2),
      "getRepeatId() returns same repeat id for objects with the same values");
  }

  function testDifferentValues() {
    const message1 = stubConsoleMessages.get("console.log('foobar', 'test')");
    const message2 = message1.set("parameters", ["funny", "monkey"]);

    isnot(getRepeatId(message1), getRepeatId(message2),
      "getRepeatId() returns different repeat ids for different values");
  }

  function testDifferentSeverities() {
    const message1 = stubConsoleMessages.get("console.log('foobar', 'test')");
    const message2 = message1.set("level", "error");

    isnot(getRepeatId(message1), getRepeatId(message2),
      "getRepeatId() returns different repeat ids for different severities");
  }

  function testFalsyValues() {
    const messageNaN = stubConsoleMessages.get("console.log(NaN)");
    const messageUnd = stubConsoleMessages.get("console.log(undefined)");
    const messageNul = stubConsoleMessages.get("console.log(null)");

    const repeatIds = new Set([
      getRepeatId(messageNaN),
      getRepeatId(messageUnd),
      getRepeatId(messageNul)]
    );
    is(repeatIds.size, 3,
      "getRepeatId() handles falsy values distinctly");
  }
});
</script>
</body>
</html>
